Split types into separate crate:
Enables using them from importing projects
