+++ zomes/trust_atom/Cargo.toml
@@ -14,6 +14,7 @@ crate-type = ["cdylib", "rlib"]
 hdk = "=0.1.1"
 rust_decimal = "1"
 serde = "1"
+trust_atom_types = { path = "../trust_atom_types" }
 trust_atom_integrity = { path = "../trust_atom_integrity" }
 
 [dev-dependencies]
